// src/app/model/access-request-approvals-pending.ts
export class AccessRequestApprovalsPending {
    id: string;
    name: string;
    requestedObjectName: string;
    requestedObjectType: string;
    requestType: string;
    requester: string;
    requestedFor: string;
    owner: string;
    requestCreated: string;
    removeDate: string;
    created: string;
    sodViolationState: string;
}